+++ geothermsim/__init__.py
@@ -5,6 +5,7 @@ from .borefield import Borefield, Network
 from .ground_heat_exchanger import UTubeHeatExchanger, CoaxialHeatExchanger
 from .thermal_resistance import Multipole
 from .simulation import LoadAggregation, Simulation, LoadHistoryReconstruction, gFunction
+from . import heat_transfer
 from . import thermal_resistance
 from . import utilities
 from . import plot