+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 * [Issue 3](https://github.com/MassimoCimmino/geothermsim/issues/3) - Introduced new initialization options for the `Path` class. The trajectory can now be represented as a spline passing through an array of positions. Functions that evaluate *g*-functions can now be jit-compiled and are compatible with automatic differentiation. This also removes `numpy` as a dependency of `geothermsim` but adds `interpax`.
+* [Issue 10](https://github.com/MassimoCimmino/geothermsim/issues/10) - Created the `heat_transfer` module. The point heat source solution is moved from the `Path` class to the new module.
 
 
 ## Version 0.1.2 (2025-05-06)